Misc navi cleanup etc.
